Move helper functions into order-utils
